+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Support for identity cmap's ([#626](https://github.com/pdfminer/pdfminer.six/pull/626))
 
+### Fixed
+- `log.info` changed to `log.debug` in six files ([#604](https://github.com/pdfminer/pdfminer.six/issues/608))
 ## [20211012]
 
 ### Added